+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - lt_handshake() function renamed to lt_session_start()
 - Unit tests organization was updated
 - Included lt_l2_api_structs.h and lt_l3_api_structs.h are automatically generated
+- SH0 pairing keypair now reflects first batch of TROPIC01 devices
 
 ### Added
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - L2 commands: Resend, Sleep and Log req
 - Example folder and test folder is reworked
 - option LT_ADD_EXAMPLES controls if example's folder code is a part of compilation
+- libtropic API: lt_get_chip_status()
 
 ### Fixed